Rename `answer` method to stop migration conflicts
The migrations that are being run at the same time as this code change are referrering to question.answer. The intent is for this to reference the old `Answer` table through association. In practice this was calling the new `#answer` method on question and failing in the case where a radio question doesn't exist in the database. Instead it will return nil (some questions don't have answers) and then try to find the `ShortTextAnswer` association.

We want this handy method but we want to avoid conflicts for now. Change the method name to be more verbose. We can rename it back once this migration has been performed.